backend/requirements.txt
fastapi
sqlalchemy 
pyodbc 
pymssql  
passlib[bcrypt]
pyjwt
google-generativeai
uvicorn